chore: change dstToken type

+++ contracts/Vault.sol
@@ -25,11 +25,11 @@ contract Vault is IVault, AggKeyNonceConsumer, GovernanceCommunityGuarded {
     event TransferNativeFailed(address payable indexed recipient, uint256 amount);
     event TransferTokenFailed(address payable indexed recipient, uint256 amount, address indexed token, bytes reason);
 
-    event SwapNative(uint32 dstChain, bytes dstAddress, uint16 dstToken, uint256 amount, address indexed sender);
+    event SwapNative(uint32 dstChain, bytes dstAddress, uint32 dstToken, uint256 amount, address indexed sender);
     event SwapToken(
         uint32 dstChain,
         bytes dstAddress,
-        uint16 dstToken,
+        uint32 dstToken,
         address srcToken,
         uint256 amount,
         address indexed sender
@@ -42,7 +42,7 @@ contract Vault is IVault, AggKeyNonceConsumer, GovernanceCommunityGuarded {
     event XCallNative(
         uint32 dstChain,
         bytes dstAddress,
-        uint16 dstToken,
+        uint32 dstToken,
         uint256 amount,
         address indexed sender,
         bytes message,
@@ -52,7 +52,7 @@ contract Vault is IVault, AggKeyNonceConsumer, GovernanceCommunityGuarded {
     event XCallToken(
         uint32 dstChain,
         bytes dstAddress,
-        uint16 dstToken,
+        uint32 dstToken,
         address srcToken,
         uint256 amount,
         address indexed sender,
@@ -353,7 +353,7 @@ contract Vault is IVault, AggKeyNonceConsumer, GovernanceCommunityGuarded {
     function xSwapNative(
         uint32 dstChain,
         bytes memory dstAddress,
-        uint16 dstToken
+        uint32 dstToken
     ) external payable override onlyNotSuspended nzUint(msg.value) {
         emit SwapNative(dstChain, dstAddress, dstToken, msg.value, msg.sender);
     }
@@ -373,7 +373,7 @@ contract Vault is IVault, AggKeyNonceConsumer, GovernanceCommunityGuarded {
     function xSwapToken(
         uint32 dstChain,
         bytes memory dstAddress,
-        uint16 dstToken,
+        uint32 dstToken,
         IERC20 srcToken,
         uint256 amount
     ) external override onlyNotSuspended nzUint(amount) {
@@ -407,7 +407,7 @@ contract Vault is IVault, AggKeyNonceConsumer, GovernanceCommunityGuarded {
     function xCallNative(
         uint32 dstChain,
         bytes calldata dstAddress,
-        uint16 dstToken,
+        uint32 dstToken,
         bytes calldata message,
         uint256 gasAmount,
         bytes calldata refundAddress
@@ -440,7 +440,7 @@ contract Vault is IVault, AggKeyNonceConsumer, GovernanceCommunityGuarded {
     function xCallToken(
         uint32 dstChain,
         bytes memory dstAddress,
-        uint16 dstToken,
+        uint32 dstToken,
         bytes calldata message,
         uint256 gasAmount,
         IERC20 srcToken,
